+++ packages/react-client-sdk-checkout-with-eth/src/components/checkoutWithEth/ViewPricingDetails.tsx
@@ -146,11 +146,21 @@ export const ViewPricingDetails = ({
             }
           }
 
-          // try switching network first if needed or supported
+          // Prompt to switch the network.
           try {
             if (chainId !== data.chainId && switchNetworkAsync) {
               console.log(`switching signer network to chainId: ${chainId}`);
               await switchNetworkAsync(data.chainId);
+              console.log(
+                `Either switched network or sent message telling user to change networks`,
+              );
+
+              // There is a known issue with the underlying state of the provider
+              // still referencing the old chain that will cause the sendTransaction call
+              // to fail.
+              // Return here to stop execution.
+              // The iframe will reload connected on the correct chain.
+              return;
             } else if (chainId !== data.chainId) {
               throw {
                 isErrorObject: true,
@@ -159,7 +169,7 @@ export const ViewPricingDetails = ({
               };
             }
           } catch (error) {
-            console.log("error switching network");
+            console.log("Error switching network.", error);
             handlePayWithCryptoError(error as Error, onError, (errorObject) => {
               if (iframeRef.current) {
                 postMessageToIframe(iframeRef.current, PAY_WITH_ETH_ERROR, {
@@ -171,9 +181,8 @@ export const ViewPricingDetails = ({
             return;
           }
 
-          // send the transaction
+          // Prompt to send funds.
           try {
-            console.log("sending funds...", data);
             const result = await sendTransactionAsync?.({
               chainId: data.chainId,
               request: {
@@ -203,7 +212,7 @@ export const ViewPricingDetails = ({
               });
             }
           } catch (error) {
-            console.log("error sending funds", error);
+            console.log("Error sending funds:", error);
             handlePayWithCryptoError(error as Error, onError, (errorObject) => {
               if (iframeRef.current) {
                 postMessageToIframe(iframeRef.current, PAY_WITH_ETH_ERROR, {