+++ packages/react-client-sdk-checkout-with-eth/src/lib/hooks/useSwitchNetwork.ts
@@ -2,11 +2,8 @@ import type { ethers } from "ethers";
 import { useCallback } from "react";
 
 import type { Chain } from "wagmi";
-import {
-  allChains,
-  SwitchChainNotSupportedError,
-  useSwitchNetwork as useSwitchNetworkWagmi,
-} from "wagmi";
+import { useSwitchNetwork as useSwitchNetworkWagmi } from "wagmi";
+import { WagmiChains } from "../../components/checkoutWithEth";
 
 export const useSwitchNetwork = ({
   signer: signer,
@@ -22,16 +19,21 @@ export const useSwitchNetwork = ({
       if (_switchNetworkAsync) {
         return await _switchNetworkAsync?.(chainId);
       } else if (signer) {
-        const chainToSwitchTo = allChains.find((x) => x.id === chainId);
+        const chainToSwitchTo = WagmiChains.find((x) => x.id === chainId);
         if (!chainToSwitchTo) {
-          throw SwitchChainNotSupportedError;
+          const error = `Error switching chain. Please switch your network to chain with chainId: ${chainId}`;
+          throw new Error(error);
         }
         if ((await signer.getChainId()) !== chainId) {
-          throw SwitchChainNotSupportedError;
+          const error = `Error switching chain. Please switch your network to ${
+            chainToSwitchTo?.name ?? `chain with chainId: ${chainId}`
+          }`;
+          throw new Error(error);
         }
         return chainToSwitchTo;
       }
-      throw SwitchChainNotSupportedError;
+      const error = `Error switching chain. Please switch your network to chain with chainId: ${chainId}`;
+      throw new Error(error);
     },
     [signer, _switchNetworkAsync],
   );