+++ packages/react-client-sdk-checkout-with-eth/src/lib/utils/handleError.ts
@@ -12,67 +12,61 @@ export function handlePayWithCryptoError(
   onError?: (code: PaperSDKError) => void,
   postToParent?: (errorObject: Omit<IErrorObject, "isErrorObject">) => void,
 ) {
-  if ("isErrorObject" in error) {
-    if (onError) {
-      onError({ code: error.title, error: new Error(error.title) });
-    }
-    if (postToParent) {
-      postToParent({ ...error });
-    }
+  // No error passed.
+  if (!error) {
+    const error = new Error("Something went wrong, contact support.");
+    onError?.({ code: PayWithCryptoErrorCode.ErrorSendingTransaction, error });
+    postToParent?.({
+      title: PayWithCryptoErrorCode.ErrorSendingTransaction,
+      description: "Something went wrong, contact support.",
+    });
+    // Custom error object.
+  } else if ("isErrorObject" in error) {
+    onError?.({ code: error.title, error: new Error(error.title) });
+    postToParent?.({ ...error });
+    // Users denied the transaction.
+  } else if (
+    error?.message?.includes("rejected") ||
+    error?.message?.includes("denied transaction")
+  ) {
+    onError?.({ code: PayWithCryptoErrorCode.TransactionCancelled, error });
+    postToParent?.({
+      title: PayWithCryptoErrorCode.TransactionCancelled,
+      description: "",
+    });
+    // Insufficient funds.
+  } else if (error?.message?.includes("insufficient funds")) {
+    onError?.({
+      code: PayWithCryptoErrorCode.InsufficientBalance,
+      error,
+    });
+    postToParent?.({
+      title: PayWithCryptoErrorCode.InsufficientBalance,
+      description: `Check your wallet's ETH balance to make sure you have enough`,
+    });
+    // Failed to switch chains
+  } else if (error?.message?.includes("Error switching chain")) {
+    onError?.({
+      code: PayWithCryptoErrorCode.ChainSwitchUnderway,
+      error,
+    });
+    postToParent?.({
+      title: PayWithCryptoErrorCode.ChainSwitchUnderway,
+      description: "Check your wallet app",
+    });
+    // Catch-all.
   } else {
-    if (
-      error.message.includes("rejected") ||
-      error.message.includes("denied transaction")
-    ) {
-      if (onError) {
-        onError({ code: PayWithCryptoErrorCode.TransactionCancelled, error });
-      }
-      if (postToParent) {
-        postToParent({
-          description: "",
-          title: PayWithCryptoErrorCode.TransactionCancelled,
-        });
-      }
-    } else if (error.message.includes("insufficient funds")) {
-      if (onError) {
-        onError({
-          code: PayWithCryptoErrorCode.InsufficientBalance,
-          error,
-        });
-      }
-      if (postToParent) {
-        postToParent({
-          description:
-            "Check your wallet's ETH balance to make sure you have enough!",
-          title: PayWithCryptoErrorCode.InsufficientBalance,
-        });
-      }
-    } else if (error.message.includes("Error switching chain")) {
-      if (onError) {
-        onError({
-          code: PayWithCryptoErrorCode.ChainSwitchUnderway,
-          error,
-        });
-      }
-      if (postToParent) {
-        postToParent({
-          description: "Check your wallet app",
-          title: PayWithCryptoErrorCode.ChainSwitchUnderway,
-        });
-      }
-    } else {
-      if (onError) {
-        onError({
-          code: PayWithCryptoErrorCode.ErrorSendingTransaction,
-          error,
-        });
-      }
-      if (postToParent) {
-        postToParent({
-          description: `${error.message}`,
-          title: PayWithCryptoErrorCode.ErrorSendingTransaction,
-        });
-      }
-    }
+    onError?.({
+      code: PayWithCryptoErrorCode.ErrorSendingTransaction,
+      error: error.message
+        ? error
+        : new Error(`Something went wrong when sending the transaction`),
+    });
+    postToParent?.({
+      title: PayWithCryptoErrorCode.ErrorSendingTransaction,
+      description:
+        `${error?.message}` ??
+        `Something went wrong when sending the transaction`,
+    });
   }
 }