+++ packages/embedded-wallet-service-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @paperxyz/embedded-wallet-service-sdk
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [cec4455]
+- Updated dependencies [cec4455]
+  - @paperxyz/sdk-common-utilities@0.1.0
+
+## 1.1.1
+
+### Patch Changes
+
+- e4d003f: chore(embedded-wallet-service-sdk): release new internal constant
+
 ## 1.1.0
 
 ### Minor Changes